Clean up API readall test
This throws everthing into a common module to avoid namespace
issues in the future along with loading everything using autoload.

Beyond that I fixed a few bugs that cropped up because the code
in the main brew repo has changed recently. Specifically, the
fact that we now have a CoreCaskTap instead of just getting
the cask tap directly.
