+++ cmd/api-readall-test.rb
@@ -4,6 +4,7 @@
 ENV["HOMEBREW_NO_INSTALL_FROM_API"] = "1"
 
 require "cli/parser"
+require_relative "../lib/hd_utils"
 
 module Homebrew
   def self.api_readall_test_args
@@ -29,8 +30,7 @@ module Homebrew
     args = api_readall_test_args.parse
 
     unless args.cask?
-      require_relative "../lib/api-readall/formula_test"
-      APIReadall::FormulaTest.run(
+      HDUtils::APIReadall::FormulaTest.run(
         quiet: args.quiet?,
         verbose: args.verbose?,
         fail_fast: args.fail_fast?,
@@ -40,8 +40,7 @@ module Homebrew
     puts if !args.formula? && !args.cask?
 
     unless args.formula?
-      require_relative "../lib/api-readall/cask_test"
-      APIReadall::CaskTest.run(
+      HDUtils::APIReadall::CaskTest.run(
         quiet: args.quiet?,
         verbose: args.verbose?,
         fail_fast: args.fail_fast?,
