+++ lib/hd-utils/api_readall.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HDUtils
+  module APIReadall
+    autoload :CaskTest, "hd-utils/api-readall/cask_test"
+    autoload :FormulaTest, "hd-utils/api-readall/formula_test"
+  end
+end
